Change credit reason to be a string
The requirements around whether credit reasons should be fixed or
dynamic are unclear, so I'm just making this a string for now.

We can change credit reason in the future, but for the moment, I want to
merge the swagger file, to minimise chance of merge conflicts for Mo
(who has complained that he's editing an out-of-date file).

See
http://jira.blinkbox.local/confluence/pages/viewpage.action?pageId=21436336&focusedCommentId=21439677#comment-21439677
and
https://git.mobcastdev.com/Agora/account-credit-service-v2/pull/5/files#r9578