Test issuer info shown to users with CSR /and/ CSM
This behaviour under test was already occuring, but by accident (namely
the ordering of my if-else in AdminApi). If the order had been the other
way round, no tests would have failed, but such behaviour would have been
undesirable. With this test, we have something that will fail if we ever
regress in that manner.