CRED-38 Remove 'debit' and 'checkBalance' vuln
Prior to this change, customers might have been able to use the same credit
twice. Suppose a customer has £10 credit, and then made two debit requests,
each of £10.

Depending on timing, both debit requests could have passed the 'does the user
have enough credit to debit £10', and then still manage to debit £20 in total.
Oops

The fix is for 'debit' and 'checkBalance' to run in the same transaction, which
will be easiest if they're in the same method.

This is an example of this type of vulnerability:
http://en.wikipedia.org/wiki/Time_of_check_to_time_of_use