Refactor: Move everything in common to admin
I need to change something that is in admin, from something in common.
This of course isn't possible.

Two solutions:
1. Move the thing that is in admin, into common
2. Move the thing that is in common, into admin

(1) will eventually be the right solution, once we actually need common
(to share code with public), but until we start (2), we won't actually
know what needs to be common, so everything should be in admin for now.

We can move stuff into common as need, when we actually understand the
needs.